completely rewrote the data layer, using neo4j

+++ src/main/java/com/raccoonfink/cruisemonkey/dao/FavoriteDao.java
@@ -2,11 +2,9 @@ package com.raccoonfink.cruisemonkey.dao;
 
 import java.util.List;
 
-import org.hibernate.Session;
-
 import com.raccoonfink.cruisemonkey.model.Favorite;
 
-public interface FavoriteDao extends Dao<Favorite,Integer> {
+public interface FavoriteDao extends Dao<Favorite,Long> {
 	public List<Favorite> findByUser(final String userName);
 	public Favorite findByUserAndEventId(final String userName, final String eventId);
 }
